[FIX]  registries: inline non-shared registry actions
The registry item ("ActionSpec") and its action's implementation were
split before.

The goal of this commit is to inline the 'ActionSpec/execute' methods that are
only used once and are well-readable when written inline.

Task ID : 3377143
