+++ src/actions/insert_actions.ts
@@ -1,11 +1,15 @@
 import { functionRegistry } from "../functions";
 import { isConsecutive, isDefined } from "../helpers";
+import { handlePasteResult } from "../helpers/ui/paste_interactive";
 import { _lt } from "../translation";
 import { ActionBuilder, ActionSpec } from "./action";
 import * as ACTIONS from "./menu_items_actions";
 
 export const insertRow: ActionSpec = {
-  name: ACTIONS.MENU_INSERT_ROWS_NAME,
+  name: (env) => {
+    const number = getRowsNumber(env);
+    return number === 1 ? _lt("Insert row") : _lt("Insert %s rows", number.toString());
+  },
   isVisible: (env) =>
     isConsecutive(env.model.getters.getActiveRows()) &&
     ACTIONS.IS_ONLY_ONE_RANGE(env) &&
@@ -14,7 +18,10 @@ export const insertRow: ActionSpec = {
 };
 
 export const rowInsertRowBefore: ActionSpec = {
-  name: ACTIONS.ROW_INSERT_ROWS_BEFORE_NAME,
+  name: (env) => {
+    const number = getRowsNumber(env);
+    return number === 1 ? _lt("Insert row above") : _lt("Insert %s rows above", number.toString());
+  },
   execute: ACTIONS.INSERT_ROWS_BEFORE_ACTION,
   isVisible: (env) =>
     isConsecutive(env.model.getters.getActiveRows()) &&
@@ -25,19 +32,34 @@ export const rowInsertRowBefore: ActionSpec = {
 
 export const topBarInsertRowsBefore: ActionSpec = {
   ...rowInsertRowBefore,
-  name: ACTIONS.MENU_INSERT_ROWS_BEFORE_NAME,
+  name: (env) => {
+    const number = getRowsNumber(env);
+    if (number === 1) {
+      return _lt("Row above");
+    }
+    return _lt("%s Rows above", number.toString());
+  },
 };
 
 export const cellInsertRowsBefore: ActionSpec = {
   ...rowInsertRowBefore,
-  name: ACTIONS.CELL_INSERT_ROWS_BEFORE_NAME,
+  name: (env) => {
+    const number = getRowsNumber(env);
+    if (number === 1) {
+      return _lt("Insert row");
+    }
+    return _lt("Insert %s rows", number.toString());
+  },
   isVisible: ACTIONS.IS_ONLY_ONE_RANGE,
   icon: "o-spreadsheet-Icon.INSERT_ROW_BEFORE",
 };
 
 export const rowInsertRowsAfter: ActionSpec = {
   execute: ACTIONS.INSERT_ROWS_AFTER_ACTION,
-  name: ACTIONS.ROW_INSERT_ROWS_AFTER_NAME,
+  name: (env) => {
+    const number = getRowsNumber(env);
+    return number === 1 ? _lt("Insert row below") : _lt("Insert %s rows below", number.toString());
+  },
   isVisible: (env) =>
     isConsecutive(env.model.getters.getActiveRows()) &&
     ACTIONS.IS_ONLY_ONE_RANGE(env) &&
@@ -47,11 +69,20 @@ export const rowInsertRowsAfter: ActionSpec = {
 
 export const topBarInsertRowsAfter: ActionSpec = {
   ...rowInsertRowsAfter,
-  name: ACTIONS.MENU_INSERT_ROWS_AFTER_NAME,
+  name: (env) => {
+    const number = getRowsNumber(env);
+    if (number === 1) {
+      return _lt("Row below");
+    }
+    return _lt("%s Rows below", number.toString());
+  },
 };
 
 export const insertCol: ActionSpec = {
-  name: ACTIONS.MENU_INSERT_COLUMNS_NAME,
+  name: (env) => {
+    const number = getColumnsNumber(env);
+    return number === 1 ? _lt("Insert column") : _lt("Insert %s columns", number.toString());
+  },
   isVisible: (env) =>
     isConsecutive(env.model.getters.getActiveCols()) &&
     ACTIONS.IS_ONLY_ONE_RANGE(env) &&
@@ -60,7 +91,12 @@ export const insertCol: ActionSpec = {
 };
 
 export const colInsertColsBefore: ActionSpec = {
-  name: ACTIONS.COLUMN_INSERT_COLUMNS_BEFORE_NAME,
+  name: (env) => {
+    const number = getColumnsNumber(env);
+    return number === 1
+      ? _lt("Insert column left")
+      : _lt("Insert %s columns left", number.toString());
+  },
   execute: ACTIONS.INSERT_COLUMNS_BEFORE_ACTION,
   isVisible: (env) =>
     isConsecutive(env.model.getters.getActiveCols()) &&
@@ -71,18 +107,35 @@ export const colInsertColsBefore: ActionSpec = {
 
 export const topBarInsertColsBefore: ActionSpec = {
   ...colInsertColsBefore,
-  name: ACTIONS.MENU_INSERT_COLUMNS_BEFORE_NAME,
+  name: (env) => {
+    const number = getColumnsNumber(env);
+    if (number === 1) {
+      return _lt("Column left");
+    }
+    return _lt("%s Columns left", number.toString());
+  },
 };
 
 export const cellInsertColsBefore: ActionSpec = {
   ...colInsertColsBefore,
-  name: ACTIONS.CELL_INSERT_COLUMNS_BEFORE_NAME,
+  name: (env) => {
+    const number = getColumnsNumber(env);
+    if (number === 1) {
+      return _lt("Insert column");
+    }
+    return _lt("Insert %s columns", number.toString());
+  },
   isVisible: ACTIONS.IS_ONLY_ONE_RANGE,
   icon: "o-spreadsheet-Icon.INSERT_COL_BEFORE",
 };
 
 export const colInsertColsAfter: ActionSpec = {
-  name: ACTIONS.COLUMN_INSERT_COLUMNS_AFTER_NAME,
+  name: (env) => {
+    const number = getColumnsNumber(env);
+    return number === 1
+      ? _lt("Insert column right")
+      : _lt("Insert %s columns right", number.toString());
+  },
   execute: ACTIONS.INSERT_COLUMNS_AFTER_ACTION,
   isVisible: (env) =>
     isConsecutive(env.model.getters.getActiveCols()) &&
@@ -93,7 +146,13 @@ export const colInsertColsAfter: ActionSpec = {
 
 export const topBarInsertColsAfter: ActionSpec = {
   ...colInsertColsAfter,
-  name: ACTIONS.MENU_INSERT_COLUMNS_AFTER_NAME,
+  name: (env) => {
+    const number = getColumnsNumber(env);
+    if (number === 1) {
+      return _lt("Column right");
+    }
+    return _lt("%s Columns right", number.toString());
+  },
   execute: ACTIONS.INSERT_COLUMNS_AFTER_ACTION,
 };
 
@@ -108,7 +167,11 @@ export const insertCell: ActionSpec = {
 
 export const insertCellShiftDown: ActionSpec = {
   name: _lt("Insert cells and shift down"),
-  execute: ACTIONS.INSERT_CELL_SHIFT_DOWN,
+  execute: (env) => {
+    const zone = env.model.getters.getSelectedZone();
+    const result = env.model.dispatch("INSERT_CELL", { zone, shiftDimension: "ROW" });
+    handlePasteResult(env, result);
+  },
   isVisible: (env) =>
     env.model.getters.getActiveRows().size === 0 && env.model.getters.getActiveCols().size === 0,
   icon: "o-spreadsheet-Icon.INSERT_CELL_SHIFT_DOWN",
@@ -116,7 +179,11 @@ export const insertCellShiftDown: ActionSpec = {
 
 export const insertCellShiftRight: ActionSpec = {
   name: _lt("Insert cells and shift right"),
-  execute: ACTIONS.INSERT_CELL_SHIFT_RIGHT,
+  execute: (env) => {
+    const zone = env.model.getters.getSelectedZone();
+    const result = env.model.dispatch("INSERT_CELL", { zone, shiftDimension: "COL" });
+    handlePasteResult(env, result);
+  },
   isVisible: (env) =>
     env.model.getters.getActiveRows().size === 0 && env.model.getters.getActiveCols().size === 0,
   icon: "o-spreadsheet-Icon.INSERT_CELL_SHIFT_RIGHT",
@@ -201,7 +268,13 @@ export const insertLink: ActionSpec = {
 
 export const insertSheet: ActionSpec = {
   name: _lt("Insert sheet"),
-  execute: ACTIONS.CREATE_SHEET_ACTION,
+  execute: (env) => {
+    const activeSheetId = env.model.getters.getActiveSheetId();
+    const position = env.model.getters.getSheetIds().indexOf(activeSheetId) + 1;
+    const sheetId = env.model.uuidGenerator.uuidv4();
+    env.model.dispatch("CREATE_SHEET", { sheetId, position });
+    env.model.dispatch("ACTIVATE_SHEET", { sheetIdFrom: activeSheetId, sheetIdTo: sheetId });
+  },
   icon: "o-spreadsheet-Icon.INSERT_SHEET",
 };
 
@@ -213,4 +286,24 @@ function createFormulaFunctions(fnNames: string[]): ActionSpec[] {
       execute: (env) => env.startCellEdition(`=${fnName}(`),
     };
   });
+}
+
+function getRowsNumber(env): number {
+  const activeRows = env.model.getters.getActiveRows();
+  if (activeRows.size) {
+    return activeRows.size;
+  } else {
+    const zone = env.model.getters.getSelectedZones()[0];
+    return zone.bottom - zone.top + 1;
+  }
+}
+
+function getColumnsNumber(env): number {
+  const activeCols = env.model.getters.getActiveCols();
+  if (activeCols.size) {
+    return activeCols.size;
+  } else {
+    const zone = env.model.getters.getSelectedZones()[0];
+    return zone.right - zone.left + 1;
+  }
 }
