+++ src/actions/menu_items_actions.ts
@@ -4,15 +4,8 @@ import {
   getSmartChartDefinition,
 } from "../helpers/figures/charts";
 import { centerFigurePosition, getMaxFigureSize } from "../helpers/figures/figure/figure";
-import { areZonesContinuous, getZoneArea, numberToLetters } from "../helpers/index";
-import { interactiveSortSelection } from "../helpers/sort";
-import { interactiveCut } from "../helpers/ui/cut_interactive";
-import { interactiveAddFilter } from "../helpers/ui/filter_interactive";
-import {
-  handlePasteResult,
-  interactivePaste,
-  interactivePasteFromOS,
-} from "../helpers/ui/paste_interactive";
+import { getZoneArea, numberToLetters } from "../helpers/index";
+import { interactivePaste, interactivePasteFromOS } from "../helpers/ui/paste_interactive";
 import { _lt } from "../translation";
 import { ClipboardMIMEType, ClipboardPasteOptions } from "../types/clipboard";
 import { Image } from "../types/image";
@@ -21,26 +14,6 @@ import { Format, SpreadsheetChildEnv, Style } from "../types/index";
 //------------------------------------------------------------------------------
 // Helpers
 //------------------------------------------------------------------------------
-
-function getColumnsNumber(env: SpreadsheetChildEnv): number {
-  const activeCols = env.model.getters.getActiveCols();
-  if (activeCols.size) {
-    return activeCols.size;
-  } else {
-    const zone = env.model.getters.getSelectedZones()[0];
-    return zone.right - zone.left + 1;
-  }
-}
-
-function getRowsNumber(env: SpreadsheetChildEnv): number {
-  const activeRows = env.model.getters.getActiveRows();
-  if (activeRows.size) {
-    return activeRows.size;
-  } else {
-    const zone = env.model.getters.getSelectedZones()[0];
-    return zone.bottom - zone.top + 1;
-  }
-}
 
 export function setFormatter(env: SpreadsheetChildEnv, format: Format) {
   env.model.dispatch("SET_FORMATTING", {
@@ -61,20 +34,6 @@ export function setStyle(env: SpreadsheetChildEnv, style: Style) {
 //------------------------------------------------------------------------------
 // Simple actions
 //------------------------------------------------------------------------------
-
-export const UNDO_ACTION = (env: SpreadsheetChildEnv) => env.model.dispatch("REQUEST_UNDO");
-
-export const REDO_ACTION = (env: SpreadsheetChildEnv) => env.model.dispatch("REQUEST_REDO");
-
-export const COPY_ACTION = async (env: SpreadsheetChildEnv) => {
-  env.model.dispatch("COPY");
-  await env.clipboard.write(env.model.getters.getClipboardContent());
-};
-
-export const CUT_ACTION = async (env: SpreadsheetChildEnv) => {
-  interactiveCut(env);
-  await env.clipboard.write(env.model.getters.getClipboardContent());
-};
 
 export const PASTE_ACTION = async (env: SpreadsheetChildEnv) => paste(env);
 export const PASTE_VALUE_ACTION = async (env: SpreadsheetChildEnv) => paste(env, "onlyValue");
@@ -113,27 +72,6 @@ async function paste(env: SpreadsheetChildEnv, pasteOption?: ClipboardPasteOptio
 }
 
 export const PASTE_FORMAT_ACTION = (env: SpreadsheetChildEnv) => paste(env, "onlyFormat");
-
-export const DELETE_CONTENT_ACTION = (env: SpreadsheetChildEnv) =>
-  env.model.dispatch("DELETE_CONTENT", {
-    sheetId: env.model.getters.getActiveSheetId(),
-    target: env.model.getters.getSelectedZones(),
-  });
-
-export const SET_FORMULA_VISIBILITY_ACTION = (env: SpreadsheetChildEnv) =>
-  env.model.dispatch("SET_FORMULA_VISIBILITY", { show: !env.model.getters.shouldShowFormulas() });
-
-export const SET_GRID_LINES_VISIBILITY_ACTION = (env: SpreadsheetChildEnv) => {
-  const sheetId = env.model.getters.getActiveSheetId();
-  env.model.dispatch("SET_GRID_LINES_VISIBILITY", {
-    sheetId,
-    areGridLinesVisible: !env.model.getters.getGridLinesVisibility(sheetId),
-  });
-};
-
-export const IS_NOT_CUT_OPERATION = (env: SpreadsheetChildEnv): boolean => {
-  return !env.model.getters.isCutOperation();
-};
 
 //------------------------------------------------------------------------------
 // Grid manipulations
@@ -287,56 +225,6 @@ export const NOT_ALL_VISIBLE_COLS_SELECTED = (env: SpreadsheetChildEnv) => {
   return env.model.getters.canRemoveHeaders(sheetId, "COL", selectedCols);
 };
 
-export const INSERT_CELL_SHIFT_DOWN = (env: SpreadsheetChildEnv) => {
-  const zone = env.model.getters.getSelectedZone();
-  const result = env.model.dispatch("INSERT_CELL", { zone, shiftDimension: "ROW" });
-  handlePasteResult(env, result);
-};
-
-export const INSERT_CELL_SHIFT_RIGHT = (env: SpreadsheetChildEnv) => {
-  const zone = env.model.getters.getSelectedZone();
-  const result = env.model.dispatch("INSERT_CELL", { zone, shiftDimension: "COL" });
-  handlePasteResult(env, result);
-};
-
-export const DELETE_CELL_SHIFT_UP = (env: SpreadsheetChildEnv) => {
-  const zone = env.model.getters.getSelectedZone();
-  const result = env.model.dispatch("DELETE_CELL", { zone, shiftDimension: "ROW" });
-  handlePasteResult(env, result);
-};
-
-export const DELETE_CELL_SHIFT_LEFT = (env: SpreadsheetChildEnv) => {
-  const zone = env.model.getters.getSelectedZone();
-  const result = env.model.dispatch("DELETE_CELL", { zone, shiftDimension: "COL" });
-  handlePasteResult(env, result);
-};
-
-export const MENU_INSERT_ROWS_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getRowsNumber(env);
-  return number === 1 ? _lt("Insert row") : _lt("Insert %s rows", number.toString());
-};
-
-export const MENU_INSERT_ROWS_BEFORE_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getRowsNumber(env);
-  if (number === 1) {
-    return _lt("Row above");
-  }
-  return _lt("%s Rows above", number.toString());
-};
-
-export const ROW_INSERT_ROWS_BEFORE_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getRowsNumber(env);
-  return number === 1 ? _lt("Insert row above") : _lt("Insert %s rows above", number.toString());
-};
-
-export const CELL_INSERT_ROWS_BEFORE_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getRowsNumber(env);
-  if (number === 1) {
-    return _lt("Insert row");
-  }
-  return _lt("Insert %s rows", number.toString());
-};
-
 export const INSERT_ROWS_BEFORE_ACTION = (env: SpreadsheetChildEnv) => {
   const activeRows = env.model.getters.getActiveRows();
   let row: number;
@@ -356,19 +244,6 @@ export const INSERT_ROWS_BEFORE_ACTION = (env: SpreadsheetChildEnv) => {
     quantity,
     dimension: "ROW",
   });
-};
-
-export const MENU_INSERT_ROWS_AFTER_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getRowsNumber(env);
-  if (number === 1) {
-    return _lt("Row below");
-  }
-  return _lt("%s Rows below", number.toString());
-};
-
-export const ROW_INSERT_ROWS_AFTER_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getRowsNumber(env);
-  return number === 1 ? _lt("Insert row below") : _lt("Insert %s rows below", number.toString());
 };
 
 export const INSERT_ROWS_AFTER_ACTION = (env: SpreadsheetChildEnv) => {
@@ -392,34 +267,6 @@ export const INSERT_ROWS_AFTER_ACTION = (env: SpreadsheetChildEnv) => {
   });
 };
 
-export const MENU_INSERT_COLUMNS_BEFORE_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getColumnsNumber(env);
-  if (number === 1) {
-    return _lt("Column left");
-  }
-  return _lt("%s Columns left", number.toString());
-};
-
-export const MENU_INSERT_COLUMNS_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getColumnsNumber(env);
-  return number === 1 ? _lt("Insert column") : _lt("Insert %s columns", number.toString());
-};
-
-export const COLUMN_INSERT_COLUMNS_BEFORE_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getColumnsNumber(env);
-  return number === 1
-    ? _lt("Insert column left")
-    : _lt("Insert %s columns left", number.toString());
-};
-
-export const CELL_INSERT_COLUMNS_BEFORE_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getColumnsNumber(env);
-  if (number === 1) {
-    return _lt("Insert column");
-  }
-  return _lt("Insert %s columns", number.toString());
-};
-
 export const INSERT_COLUMNS_BEFORE_ACTION = (env: SpreadsheetChildEnv) => {
   const activeCols = env.model.getters.getActiveCols();
   let column: number;
@@ -439,21 +286,6 @@ export const INSERT_COLUMNS_BEFORE_ACTION = (env: SpreadsheetChildEnv) => {
     base: column,
     quantity,
   });
-};
-
-export const MENU_INSERT_COLUMNS_AFTER_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getColumnsNumber(env);
-  if (number === 1) {
-    return _lt("Column right");
-  }
-  return _lt("%s Columns right", number.toString());
-};
-
-export const COLUMN_INSERT_COLUMNS_AFTER_NAME = (env: SpreadsheetChildEnv) => {
-  const number = getColumnsNumber(env);
-  return number === 1
-    ? _lt("Insert column right")
-    : _lt("Insert %s columns right", number.toString());
 };
 
 export const INSERT_COLUMNS_AFTER_ACTION = (env: SpreadsheetChildEnv) => {
@@ -494,33 +326,6 @@ export const HIDE_COLUMNS_NAME = (env: SpreadsheetChildEnv) => {
   }
 };
 
-export const HIDE_COLUMNS_ACTION = (env: SpreadsheetChildEnv) => {
-  const columns = env.model.getters.getElementsFromSelection("COL");
-  env.model.dispatch("HIDE_COLUMNS_ROWS", {
-    sheetId: env.model.getters.getActiveSheetId(),
-    dimension: "COL",
-    elements: columns,
-  });
-};
-
-export const UNHIDE_ALL_COLUMNS_ACTION = (env: SpreadsheetChildEnv) => {
-  const sheetId = env.model.getters.getActiveSheetId();
-  env.model.dispatch("UNHIDE_COLUMNS_ROWS", {
-    sheetId,
-    dimension: "COL",
-    elements: Array.from(Array(env.model.getters.getNumberCols(sheetId)).keys()),
-  });
-};
-
-export const UNHIDE_COLUMNS_ACTION = (env: SpreadsheetChildEnv) => {
-  const columns = env.model.getters.getElementsFromSelection("COL");
-  env.model.dispatch("UNHIDE_COLUMNS_ROWS", {
-    sheetId: env.model.getters.getActiveSheetId(),
-    dimension: "COL",
-    elements: columns,
-  });
-};
-
 export const HIDE_ROWS_NAME = (env: SpreadsheetChildEnv) => {
   const rows = env.model.getters.getElementsFromSelection("ROW");
   let first = rows[0];
@@ -532,45 +337,6 @@ export const HIDE_ROWS_NAME = (env: SpreadsheetChildEnv) => {
   } else {
     return _lt("Hide rows");
   }
-};
-
-export const HIDE_ROWS_ACTION = (env: SpreadsheetChildEnv) => {
-  const rows = env.model.getters.getElementsFromSelection("ROW");
-  env.model.dispatch("HIDE_COLUMNS_ROWS", {
-    sheetId: env.model.getters.getActiveSheetId(),
-    dimension: "ROW",
-    elements: rows,
-  });
-};
-
-export const UNHIDE_ALL_ROWS_ACTION = (env: SpreadsheetChildEnv) => {
-  const sheetId = env.model.getters.getActiveSheetId();
-  env.model.dispatch("UNHIDE_COLUMNS_ROWS", {
-    sheetId,
-    dimension: "ROW",
-    elements: Array.from(Array(env.model.getters.getNumberRows(sheetId)).keys()),
-  });
-};
-
-export const UNHIDE_ROWS_ACTION = (env: SpreadsheetChildEnv) => {
-  const columns = env.model.getters.getElementsFromSelection("ROW");
-  env.model.dispatch("UNHIDE_COLUMNS_ROWS", {
-    sheetId: env.model.getters.getActiveSheetId(),
-    dimension: "ROW",
-    elements: columns,
-  });
-};
-
-//------------------------------------------------------------------------------
-// Sheets
-//------------------------------------------------------------------------------
-
-export const CREATE_SHEET_ACTION = (env: SpreadsheetChildEnv) => {
-  const activeSheetId = env.model.getters.getActiveSheetId();
-  const position = env.model.getters.getSheetIds().indexOf(activeSheetId) + 1;
-  const sheetId = env.model.uuidGenerator.uuidv4();
-  env.model.dispatch("CREATE_SHEET", { sheetId, position });
-  env.model.dispatch("ACTIVATE_SHEET", { sheetIdFrom: activeSheetId, sheetIdTo: sheetId });
 };
 
 //------------------------------------------------------------------------------
@@ -638,60 +404,13 @@ export const CREATE_IMAGE = async (env: SpreadsheetChildEnv) => {
 // Style/Format
 //------------------------------------------------------------------------------
 
-export const FORMAT_AUTOMATIC_ACTION = (env: SpreadsheetChildEnv) => setFormatter(env, "");
-
-export const FORMAT_NUMBER_ACTION = (env: SpreadsheetChildEnv) => setFormatter(env, "#,##0.00");
-
 export const FORMAT_PERCENT_ACTION = (env: SpreadsheetChildEnv) => setFormatter(env, "0.00%");
-
-export const FORMAT_CURRENCY_ACTION = (env: SpreadsheetChildEnv) =>
-  setFormatter(env, "[$$]#,##0.00");
-
-export const FORMAT_CURRENCY_ROUNDED_ACTION = (env: SpreadsheetChildEnv) =>
-  setFormatter(env, "[$$]#,##0");
-
-export const FORMAT_DATE_ACTION = (env: SpreadsheetChildEnv) => setFormatter(env, "m/d/yyyy");
-
-export const FORMAT_TIME_ACTION = (env: SpreadsheetChildEnv) => setFormatter(env, "hh:mm:ss a");
-
-export const FORMAT_DATE_TIME_ACTION = (env: SpreadsheetChildEnv) =>
-  setFormatter(env, "m/d/yyyy hh:mm:ss");
-
-export const FORMAT_DURATION_ACTION = (env: SpreadsheetChildEnv) => setFormatter(env, "hhhh:mm:ss");
-
-export const FORMAT_BOLD_ACTION = (env: SpreadsheetChildEnv) =>
-  setStyle(env, { bold: !env.model.getters.getCurrentStyle().bold });
-
-export const FORMAT_ITALIC_ACTION = (env: SpreadsheetChildEnv) =>
-  setStyle(env, { italic: !env.model.getters.getCurrentStyle().italic });
-
-export const FORMAT_STRIKETHROUGH_ACTION = (env: SpreadsheetChildEnv) =>
-  setStyle(env, { strikethrough: !env.model.getters.getCurrentStyle().strikethrough });
-
-export const FORMAT_UNDERLINE_ACTION = (env: SpreadsheetChildEnv) =>
-  setStyle(env, { underline: !env.model.getters.getCurrentStyle().underline });
-
-export const FORMAT_CLEARFORMAT_ACTION = (env: SpreadsheetChildEnv) => {
-  env.model.dispatch("CLEAR_FORMATTING", {
-    sheetId: env.model.getters.getActiveSheetId(),
-    target: env.model.getters.getSelectedZones(),
-  });
-};
 
 //------------------------------------------------------------------------------
 // Side panel
 //------------------------------------------------------------------------------
-
 export const OPEN_CF_SIDEPANEL_ACTION = (env: SpreadsheetChildEnv) => {
   env.openSidePanel("ConditionalFormatting", { selection: env.model.getters.getSelectedZones() });
-};
-
-export const OPEN_FAR_SIDEPANEL_ACTION = (env: SpreadsheetChildEnv) => {
-  env.openSidePanel("FindAndReplace", {});
-};
-
-export const OPEN_CUSTOM_CURRENCY_SIDEPANEL_ACTION = (env: SpreadsheetChildEnv) => {
-  env.openSidePanel("CustomCurrency", {});
 };
 
 export const INSERT_LINK = (env: SpreadsheetChildEnv) => {
@@ -703,46 +422,15 @@ export const INSERT_LINK = (env: SpreadsheetChildEnv) => {
 // Filters action
 //------------------------------------------------------------------------------
 
-export const FILTERS_CREATE_FILTER_TABLE = (env: SpreadsheetChildEnv) => {
-  const sheetId = env.model.getters.getActiveSheetId();
-  const selection = env.model.getters.getSelection().zones;
-  interactiveAddFilter(env, sheetId, selection);
-};
-
-export const FILTERS_REMOVE_FILTER_TABLE = (env: SpreadsheetChildEnv) => {
-  const sheetId = env.model.getters.getActiveSheetId();
-  env.model.dispatch("REMOVE_FILTER_TABLE", {
-    sheetId,
-    target: env.model.getters.getSelectedZones(),
-  });
-};
-
 export const SELECTION_CONTAINS_FILTER = (env: SpreadsheetChildEnv): boolean => {
   const sheetId = env.model.getters.getActiveSheetId();
   const selectedZones = env.model.getters.getSelectedZones();
   return env.model.getters.doesZonesContainFilter(sheetId, selectedZones);
 };
 
-export const SELECTION_IS_CONTINUOUS = (env: SpreadsheetChildEnv): boolean => {
-  const selectedZones = env.model.getters.getSelectedZones();
-  return areZonesContinuous(...selectedZones);
-};
-
 //------------------------------------------------------------------------------
 // Sorting action
 //------------------------------------------------------------------------------
-
-export const SORT_CELLS_ASCENDING = (env: SpreadsheetChildEnv) => {
-  const { anchor, zones } = env.model.getters.getSelection();
-  const sheetId = env.model.getters.getActiveSheetId();
-  interactiveSortSelection(env, sheetId, anchor.cell, zones[0], "ascending");
-};
-
-export const SORT_CELLS_DESCENDING = (env: SpreadsheetChildEnv) => {
-  const { anchor, zones } = env.model.getters.getSelection();
-  const sheetId = env.model.getters.getActiveSheetId();
-  interactiveSortSelection(env, sheetId, anchor.cell, zones[0], "descending");
-};
 
 export const IS_ONLY_ONE_RANGE = (env: SpreadsheetChildEnv): boolean => {
   return env.model.getters.getSelectedZones().length === 1;
